[Feat] Add magic methods exclustion logic
